refactor: Removed unused libs and files
