+++ components/ui/alert-dialog.tsx
@@ -1,7 +1,7 @@
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog";
 import * as React from "react";
 
-import { buttonVariants } from "../../components/ui/button";
+import { buttonVariants } from "./button";
 import { cn } from "../../lib/utils";
 
 const AlertDialog = AlertDialogPrimitive.Root;
