+++ components/ui/command.tsx
@@ -3,7 +3,7 @@ import { MagnifyingGlassIcon } from "@radix-ui/react-icons";
 import { Command as CommandPrimitive } from "cmdk";
 import * as React from "react";
 
-import { Dialog, DialogContent } from "../../components/ui/dialog";
+import { Dialog, DialogContent } from "./dialog";
 import { cn } from "../../lib/utils";
 
 const Command = React.forwardRef<
